Created special reader - This has separate contents sidebar, and back button
